// apps/frontend/src/api/models/updateCandidatePersonalInfoRequest.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCandidatePersonalInfoRequestFirstName } from './updateCandidatePersonalInfoRequestFirstName';
import type { UpdateCandidatePersonalInfoRequestLastName } from './updateCandidatePersonalInfoRequestLastName';
import type { UpdateCandidatePersonalInfoRequestPhone } from './updateCandidatePersonalInfoRequestPhone';

export interface UpdateCandidatePersonalInfoRequest {
  first_name?: UpdateCandidatePersonalInfoRequestFirstName;
  last_name?: UpdateCandidatePersonalInfoRequestLastName;
  phone?: UpdateCandidatePersonalInfoRequestPhone;
}